requirements.txt
streamlit>=1.46.0
duckdb>=1.3.0
pyyaml>=6.0
streamlit-authenticator>=0.4.0
openpyxl>=3.1.0
pandas>=2.0.0
plotly>=5.0.0
psutil>=5.9.0